Classes inside 'uk.ac.ebi.eva.pipeline.steps.tasklets' moved to the parent package
