+++ src/main/java/uk/ac/ebi/eva/pipeline/jobs/steps/VariantsLoad.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.pipeline.jobs.steps.tasklets;
+package uk.ac.ebi.eva.pipeline.jobs.steps;
 
 import org.opencb.datastore.core.ObjectMap;
 import org.opencb.opencga.storage.core.StorageManagerFactory;
