+++ src/test/java/uk/ac/ebi/eva/pipeline/jobs/VariantConfigurationWorkflowTest.java
@@ -18,9 +18,9 @@ package uk.ac.ebi.eva.pipeline.jobs;
 
 import uk.ac.ebi.eva.pipeline.configuration.VariantJobsArgs;
 import uk.ac.ebi.eva.pipeline.configuration.VariantWorkflowConfig;
+import uk.ac.ebi.eva.pipeline.jobs.steps.VariantsAnnotCreate;
 import uk.ac.ebi.eva.pipeline.jobs.steps.VariantsAnnotGenerateInput;
 import uk.ac.ebi.eva.pipeline.jobs.steps.VariantsAnnotLoad;
-import uk.ac.ebi.eva.pipeline.jobs.steps.tasklets.VariantsAnnotCreate;
 
 import java.io.File;
 import java.nio.file.Paths;
